+++ modules/dxp/apps/commerce/commerce-api/src/main/java/com/liferay/commerce/service/CommerceOrderItemLocalService.java
@@ -39,6 +39,8 @@ import com.liferay.portal.kernel.util.OrderByComparator;
 
 import java.io.Serializable;
 
+import java.math.BigDecimal;
+
 import java.util.List;
 
 /**
@@ -77,8 +79,8 @@ public interface CommerceOrderItemLocalService extends BaseLocalService,
 	@Indexable(type = IndexableType.REINDEX)
 	public CommerceOrderItem addCommerceOrderItem(long commerceOrderId,
 		long cpInstanceId, int quantity, int shippedQuantity,
-		java.lang.String json, java.lang.Double price,
-		ServiceContext serviceContext) throws PortalException;
+		java.lang.String json, BigDecimal price, ServiceContext serviceContext)
+		throws PortalException;
 
 	/**
 	* Creates a new commerce order item with the primary key. Does not add the commerce order item to the database.
@@ -292,6 +294,6 @@ public interface CommerceOrderItemLocalService extends BaseLocalService,
 
 	@Indexable(type = IndexableType.REINDEX)
 	public CommerceOrderItem updateCommerceOrderItem(long commerceOrderItemId,
-		int quantity, java.lang.String json, double price)
+		int quantity, java.lang.String json, BigDecimal price)
 		throws PortalException;
 }
