+++ modules/dxp/apps/commerce/commerce-api/src/main/java/com/liferay/commerce/service/CommerceTierPriceEntryLocalService.java
@@ -46,6 +46,8 @@ import com.liferay.portal.kernel.util.OrderByComparator;
 
 import java.io.Serializable;
 
+import java.math.BigDecimal;
+
 import java.util.List;
 
 /**
@@ -83,7 +85,7 @@ public interface CommerceTierPriceEntryLocalService extends BaseLocalService,
 
 	@Indexable(type = IndexableType.REINDEX)
 	public CommerceTierPriceEntry addCommerceTierPriceEntry(
-		long commercePriceEntryId, double price, int minQuantity,
+		long commercePriceEntryId, BigDecimal price, int minQuantity,
 		ServiceContext serviceContext) throws PortalException;
 
 	/**
@@ -333,6 +335,6 @@ public interface CommerceTierPriceEntryLocalService extends BaseLocalService,
 
 	@Indexable(type = IndexableType.REINDEX)
 	public CommerceTierPriceEntry updateCommerceTierPriceEntry(
-		long commerceTierPriceEntryId, double price, int minQuantity,
+		long commerceTierPriceEntryId, BigDecimal price, int minQuantity,
 		ServiceContext serviceContext) throws PortalException;
 }
